requirements.txt
pandas
numpy
scikit-learn
lightgbm
prophet
matplotlib
seaborn
plotly
fastapi
uvicorn[standard]
streamlit
joblib
pyyaml
loguru
pytest
